fix: treeView not found
