+++ feature/feature-treeview/src/main/java/org/gampiot/robok/feature/treeview/file/example/FileNode.java
@@ -1,0 +1,11 @@
+package org.gampiot.robok.feature.treeview.file.example;
+
+public class FileNode {
+      String name;
+      boolean isDirectory;
+      
+      public FileNode(String name, boolean isDirectory) {
+           this.name = name;
+           this.isDirectory = isDirectory;
+      }
+}
